+++ src/fractal_client/parser.py
@@ -428,7 +428,6 @@ task_edit_parser.add_argument(
         "(only accepted in combination with `--name`)."
     ),
 )
-task_edit_parser.add_argument("--new-version", help="New task version.")
 task_edit_parser.add_argument(
     "--command-non-parallel", help="New task non parallel command."
 )
@@ -679,6 +678,33 @@ workflow_export_parser.add_argument(
     help="Path to the JSON file where the workflow will be exported.",
     required=True,
 )
+
+
+# workflow import-from-template
+workflow_import_from_template_parser = workflow_subparsers.add_parser(
+    "import-from-template",
+    description="Import workflow to project from template.",
+    allow_abbrev=False,
+)
+workflow_import_from_template_parser.add_argument(
+    "project_id",
+    type=int,
+    help="ID of the project where the workflow will be imported.",
+)
+workflow_import_from_template_parser.add_argument(
+    "template_id",
+    type=int,
+    help="ID of the template from which the workflow will be imported.",
+)
+workflow_import_from_template_parser.add_argument(
+    "--name",
+    type=str,
+    help=(
+        "Name of the new workflow (if set, overrides the one in the template)."
+    ),
+    required=False,
+)
+
 
 # JOB GROUP
 
@@ -1040,3 +1066,84 @@ profile_new_parser = profile_subparsers.add_parser(
 )
 profile_new_parser.add_argument("resource_id", help="TBD")
 profile_new_parser.add_argument("json_file", help="TBD")
+
+
+# TEMPLATE GROUP
+template_parser = subparsers_main.add_parser(
+    "template",
+    description="Template commands.",
+    allow_abbrev=False,
+)
+template_subparsers = template_parser.add_subparsers(
+    title="Valid sub-commands", dest="subcmd", required=True
+)
+
+# template show
+template_show_parser = template_subparsers.add_parser(
+    "show",
+    description="Show single template.",
+    allow_abbrev=False,
+)
+template_show_parser.add_argument(
+    "template_id", help="ID of the template to show.", type=int
+)
+
+# template new
+template_new_parser = template_subparsers.add_parser(
+    "new",
+    description="Create new template.",
+    allow_abbrev=False,
+)
+
+template_new_from_workflow_or_import = (
+    template_new_parser.add_mutually_exclusive_group(required=True)
+)
+template_new_from_workflow_or_import.add_argument(
+    "--workflow-id",
+    help="ID of the workflow from which the new template will be built.",
+    type=int,
+)
+template_new_from_workflow_or_import.add_argument(
+    "--json-file",
+    help="Path to a JSON file with the template to be imported.",
+)
+
+template_new_parser.add_argument(
+    "--name", help="New template name.", required=False
+)
+template_new_parser.add_argument(
+    "--version", help="New template version.", required=False
+)
+template_new_parser.add_argument(
+    "--user-group-id",
+    help=(
+        "ID of user group which should be granted access to the new template."
+    ),
+    required=False,
+)
+
+# template delete
+template_delete_parser = template_subparsers.add_parser(
+    "delete",
+    description="Delete single template.",
+    allow_abbrev=False,
+)
+template_delete_parser.add_argument(
+    "template_id", help="ID of the template to delete.", type=int
+)
+
+
+# template export
+template_export_parser = template_subparsers.add_parser(
+    "export",
+    description="Export single template.",
+    allow_abbrev=False,
+)
+template_export_parser.add_argument(
+    "template_id",
+    help="ID of the template to export.",
+)
+template_export_parser.add_argument(
+    "json_file",
+    help="Path where to export the template.",
+)