+++ tests/test_workflow.py
@@ -483,17 +483,22 @@ def test_workflow_import(
     testdata_path,
     new_name,
     tmp_path,
+    capsys,
 ):
     name1 = new_name()
     res = invoke(
-        f"task new --command-parallel pwd --command-non-parallel pwd {name1}"
+        "task new "
+        "--command-parallel pwd --command-non-parallel pwd --version 1 "
+        f"{name1}"
     )
     debug(res.data)
     assert res.retcode == 0
 
     name2 = new_name()
     res = invoke(
-        f"task new --command-parallel pwd --command-non-parallel pwd {name2}"
+        "task new "
+        "--command-parallel pwd --command-non-parallel pwd --version 1 "
+        f"{name2}"
     )
     debug(res.data)
     assert res.retcode == 0
@@ -528,7 +533,6 @@ def test_workflow_import(
     res = invoke(f"workflow show {project_id} {workflow_id}")
     debug(res.retcode, res.data)
     assert res.retcode == 0
-    res.data["task_list"][-1]["warning"] = None
     assert res.data == imported_workflow
 
     # import workflow into project, with --batch
@@ -549,11 +553,26 @@ def test_workflow_import(
     # import workflow into project, with --workflow-name
     res = invoke(
         f"workflow import --project-id {project_id} --json-file {filename} "
-        f" --workflow-name MyWorkflow-V2-xxx"
+        " --workflow-name MyWorkflow-V2-xxx"
     )
     debug(res.data)
     assert res.retcode == 0
     assert res.data["name"] == "MyWorkflow-V2-xxx"
+
+    worfklow_to_import["task_list"][0]["task"]["version"] = "2"
+    filename = tmp_path / "workflow3"
+    with filename.open("w") as f:
+        json.dump(worfklow_to_import, f)
+    with pytest.raises(SystemExit):
+        invoke(
+            f"workflow import --project-id {project_id} --json-file {filename}"
+            " --workflow-name MyWorkflow-V2-yyy"
+        )
+    captured = capsys.readouterr()
+    assert (
+        f"Task '{name2}' (package '{name2}', version '2') not available. "
+        "Available versions: ['1']."
+    ) in captured.out
 
 
 def test_workflow_export(